Remove the xmir helpers.
Get rid of the helpers, and internal references thereto, of xmir, as we
are moving to using Xwayland for legacy app compatibility, and Mir provides
a Wayland compositor which clients can talk to directly now. We no longer need
to default to running things under Xmir. There are still some references, to
avoid breaking the ABI and API, but when we can, it should allow to simplify
the code a bit further. This is a necessary step toward making Libertine an
optional build dependency (as libertine only works with Ubuntu chroots).