desktop-hook: allow application icon to be without extension
The .desktop file specification says that the icon is either an absolute
path, or an icon name to be looked up according to the icon theme
specification (that is, it's extensionless). Moreover, the AppImage
tools emit a warning if the icon is specified as a relative path but
contains an extension.

By adding a lookup for ".png" and ".svg" extension, we allow the same
desktop file to be used for multi-platform applications.