Do not wait for handshake before triggering upstart
We have no need to wait for handshake as this has no benefit at all

The hanshake exists to let qtmir know a application is starting, but
this is not needed to be sendt before the exec of the app has begun. if
the application is started by qtmir, it will know immediately that an
app is starting so this signal is unused by qtmir. This handshake
sometimes takes 0.5s as it does many roundtrips and we simply dont need
to wait that long before exec an application.